feat: 장바구니 관련 상태 관리 훅 제거 및 컴포넌트 통합
