+++ src/types/cartItem.ts
@@ -1,0 +1,13 @@
+type ProductTypes = {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  quantity: number;
+};
+
+export type CartItemTypes = {
+  id: number;
+  quantity: 1;
+  product: ProductTypes;
+};
